// src/modules/sto/agent/enum.ts
export enum Messages {

    ALREADY_EXIST = 'wallet already exist.',
    SECCUSSFULY = 'added seccussfuly.',
    UPDATED = 'updated successfully.',
    GET_COIN_LIST = 'get all coins list',
    NOT_FOUND = 'user not found',
    FETCH_QUESTION = 'Fetch Question Listing Successfully.',
    CATEGORY_FOUND = 'Category found Successfully.',
    DATA_FOUND = 'Data found Successfully.',
    Data_NOT_FOUND = 'data not found.',

    LIST_EMPTY = "Listing Empty.",
    SOMETHING_WRONG = "Something went wrong"
    // NOT_FOUND = 'user not found',
    // NOT_FOUND = 'user not found',
    // NOT_FOUND = 'user not found',
    // NOT_FOUND = 'user not found'

}

export enum transactionStatus {
    STATUS_COMPLETE = "complete",
    blockchain_status_FAILED = "FAILED",
    CONFIRMED = "CONFIRMED",
    TOKEN_MINT_STATUS_COMPLETE = "complete"

}